bufix: fix rope bugs close #54

+++ llm/models/mg_models/llama/positional_embeddings.py
@@ -20,7 +20,7 @@ class RotaryEmbedding(torch.nn.Module):
         if self.max_seq_len_cached is None or (seq_len > self.max_seq_len_cached):
             self.max_seq_len_cached = seq_len
             # follow transformers LlamaLinearScalingRotaryEmbedding implement
-            t = torch.arange(self.max_seq_len_cached, device=x.device, dtype=self.inv_freq.dtype)
+            t = torch.arange(self.max_seq_len_cached, device=x.device, dtype=torch.float32)
             t = t / self.scale_factor
             freqs = torch.einsum('i,j->ij', t, self.inv_freq)
             # Different from paper, but it uses a different permutation in order to obtain the same calculation
